refactor: `Notification:Setting Permission Status` event in the notifications module
