+++ app/src/main/java/org/openedx/app/di/ScreenModule.kt
@@ -92,7 +92,7 @@ val screenModule = module {
             get()
         )
     }
-    viewModel { MainViewModel(get(), get(), get(), get()) }
+    viewModel { MainViewModel(get(), get(), get(), get(), get()) }
 
     factory { AuthRepository(get(), get(), get()) }
     factory { AuthInteractor(get()) }
