+++ course/src/main/java/org/openedx/course/presentation/container/CourseContainerViewModel.kt
@@ -775,12 +775,11 @@ class CourseContainerViewModel(
 
     fun requestNotificationPermission(
         activity: FragmentActivity,
-        permissionLauncher: ActivityResultLauncher<String>
+        permissionLauncher: ActivityResultLauncher<String>,
     ) {
         pushManager.requestNotificationPermission(
             activity = activity,
             permissionLauncher = permissionLauncher,
-            onRationaleShown = {}
         )
     }
 }
