+++ notifications/src/main/java/org/openedx/notifications/PushManager.kt
@@ -18,6 +18,7 @@ import org.openedx.notifications.domain.model.NotificationsPrimerConfiguration
 import org.openedx.notifications.presentation.NotificationsAnalytics
 import org.openedx.notifications.presentation.NotificationsAnalyticsEvent
 import org.openedx.notifications.presentation.NotificationsAnalyticsKey
+import org.openedx.notifications.presentation.PermissionStatus
 import org.openedx.notifications.presentation.primer.NotificationsPrimerDialogFragment
 import org.openedx.notifications.utils.PermissionUtils
 import java.util.Date
@@ -111,13 +112,30 @@ class PushManager(
     override fun requestNotificationPermission(
         activity: Activity,
         permissionLauncher: ActivityResultLauncher<String>,
-        onRationaleShown: () -> Unit
+        onSystemDialogShown: () -> Unit,
+        onRationaleShown: () -> Unit,
     ) {
         PermissionUtils.requestNotificationPermission(
             activity = activity,
             permissionLauncher = permissionLauncher,
+            onSystemDialogShown = onSystemDialogShown,
             onRationaleShown = onRationaleShown,
         )
+    }
+
+    override fun logNotificationPermissionStatusEvent(context: Context) {
+        val event = NotificationsAnalyticsEvent.NOTIFICATION_PERMISSION_STATUS
+        val permissionStatus =
+            if (NotificationManagerCompat.from(context).areNotificationsEnabled()) {
+                PermissionStatus.AUTHORIZED
+            } else {
+                PermissionStatus.DENIED
+            }
+        analytics.logEvent(event.eventName, buildMap {
+            put(NotificationsAnalyticsKey.NAME.key, event.biValue)
+            put(NotificationsAnalyticsKey.STATUS.key, permissionStatus.status)
+            put(NotificationsAnalyticsKey.CATEGORY.key, NotificationsAnalyticsKey.NOTIFICATIONS.key)
+        })
     }
 
     override fun logNotificationBellClickedEvent(hasUnreadNotifications: Boolean) {
